Delete README.md
Problem Statement -
The loan default dataset has 8 variables and 850 records, each record being loan
default status for each customer. Each Applicant was rated as “Defaulted” or
“Not-Defaulted”. New applicants for loan application can also be evaluated on
these 8 predictor variables and classified as a default or non-default based on
predictor variables. We need to identify which customer is "defaulted" and "Not-defaulted"